+++ proxy/src/telemetry/metrics/prometheus/mod.rs
@@ -49,7 +49,7 @@ use super::latency::{BUCKET_BOUNDS, Histogram};
 
 mod labels;
 use self::labels::{RequestLabels, ResponseLabels};
-pub use self::labels::Labeled;
+pub use self::labels::{DstLabels, Labeled};
 
 #[derive(Debug, Clone)]
 struct Metrics {
